# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Property": {
				"properties": {
					"name": {
						"type": "string"
					},
					"mapping": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"nullable": {
						"type": "boolean"
					},
					"primaryKey": {
						"type": "boolean"
					},
					"autoincrement": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelationTo": {
				"properties": {
					"entity": {
						"type": "string"
					},
					"property": {
						"type": "string"
					}
				},
				"required": [
					"entity",
					"property"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Relation": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"$ref": "#/components/schemas/RelationTo"
					}
				},
				"required": [
					"name",
					"type",
					"from",
					"to"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Entity": {
				"properties": {
					"name": {
						"type": "string"
					},
					"mapping": {
						"type": "string"
					},
					"properties": {
						"items": {
							"$ref": "#/components/schemas/Property"
						},
						"type": "array"
					},
					"relations": {
						"items": {
							"$ref": "#/components/schemas/Relation"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"properties",
					"relations"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Enum": {
				"properties": {
					"name": {
						"type": "string"
					},
					"values": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"name",
					"values"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Schema": {
				"properties": {
					"name": {
						"type": "string"
					},
					"entities": {
						"items": {
							"$ref": "#/components/schemas/Entity"
						},
						"type": "array"
					},
					"enums": {
						"items": {
							"$ref": "#/components/schemas/Enum"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"entities",
					"enums"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompileRequest": {
				"properties": {
					"expression": {
						"type": "string"
					},
					"schema": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"variant": {
						"type": "string"
					}
				},
				"required": [
					"expression",
					"schema",
					"language",
					"variant"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RunRequest": {
				"properties": {
					"expression": {
						"type": "string"
					},
					"connection": {
						"type": "string"
					},
					"context": {}
				},
				"required": [
					"expression",
					"connection"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "js-orm",
		"version": "0.0.1",
		"description": "ORM",
		"license": {
			"name": "Apache-2.0 License"
		},
		"contact": {
			"name": "Flavio Lionel Rita ",
			"email": "flaviolrita@hotmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/schema": {
			"post": {
				"operationId": "Apply",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Schema"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Schema"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/schema/{name}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Schema"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/expression/compile": {
			"post": {
				"operationId": "Compile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompileRequest"
							}
						}
					}
				}
			}
		},
		"/expression/query": {
			"post": {
				"operationId": "Query",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompileRequest"
							}
						}
					}
				}
			}
		},
		"/expression/schema": {
			"post": {
				"operationId": "Schema",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompileRequest"
							}
						}
					}
				}
			}
		},
		"/expression/run": {
			"post": {
				"operationId": "Run",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RunRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}